requirements.txt
# Core dependencies for agentRadar/core
pyyaml>=6.0
requests>=2.31.0
python-dateutil>=2.8.0
click>=8.1.0

# GitHub API
PyGithub>=2.1.0

# Newsletter / digest
# (Buttondown API is called via requests — no SDK needed)

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
responses>=0.23.0   # Mock HTTP requests in tests
